+++ src/components/EzModal/__tests__/EzModal.test.js
@@ -1,43 +1,82 @@
 import React from 'react';
-import {fireEvent, cleanup} from 'react-testing-library';
-import 'jest-dom/extend-expect';
+import ReactModal from 'react-modal';
 import EzModal from '../EzModal';
 import {standard} from '../../../themes';
+import EzButton from '../../EzButton';
 
-afterEach(cleanup);
+// We need to mock out the portal functionality as react test renderer does not
+// support portals currently :( . our mock forces portal s to just behave like any
+// other node
+jest.mock('react-dom', () => ({
+  createPortal: node => node,
+}));
 
 describe('EzModal', () => {
-  beforeEach(() => {
-    // suppress import style warning from Reach Modal
-    jest.spyOn(CSSStyleDeclaration.prototype, 'getPropertyValue').mockReturnValue('1');
+  let appElement;
+
+  beforeAll(() => {
+    // We need to create an arbitrary element for the modal to target as the app element
+    // React modal does not see the element if you mount it via enzyme with the modal component
+    // so we set up the element directly on the document
+    appElement = document.createElement('div');
+    ReactModal.setAppElement(appElement);
+  });
+
+  afterAll(() => {
+    document.removeChild(appElement);
   });
 
   it('should render with default styles', () => {
-    const {baseElement} = fullRender(
-      <EzModal isOpen submitLabel="Submit" dismissLabel="Dismiss" headerText="Header">
+    const actual = create(
+      <EzModal
+        isOpen
+        submitLabel="Submit"
+        dismissLabel="Dismiss"
+        headerText="Header"
+        theme={standard}
+        appElement={null}
+      >
         children
       </EzModal>
     );
-    expect(baseElement).toMatchSnapshot();
+    expect(actual).toMatchSnapshot();
   });
 
   it('does not render the children if not open', () => {
-    const {queryByText} = fullRender(
-      <EzModal isOpen={false} dismissLabel="dismiss" headerText="header">
-        <div>foo</div>
+    const component = mount(
+      <EzModal isOpen={false} appElement={null} dismissLabel="dismiss" headerText="header">
+        <div data-test-id="test">foo</div>
       </EzModal>
     );
-    expect(queryByText('foo')).toBeNull();
+    expect(component.find('[data-test-id="test"]')).toHaveLength(0);
+  });
+
+  it('calls react modals set app element on mount', () => {
+    const spy = jest.spyOn(ReactModal, 'setAppElement').mockImplementation(() => {});
+    appElement = 'myElement';
+    mount(
+      <EzModal
+        isOpen
+        appElement={appElement}
+        submitLabel="Submit"
+        dismissLabel="Dismiss"
+        headerText="Header"
+      >
+        children
+      </EzModal>
+    );
+    expect(spy).toHaveBeenCalledWith(appElement);
   });
 
   it('calls submit handler when submit button is clicked', () => {
     const clickSpy = jest.fn();
     const submitLabel = 'submit';
-    const {getByText} = fullRender(
+    const component = mount(
       <EzModal
         isOpen
         submitLabel={submitLabel}
         onSubmit={clickSpy}
+        appElement={null}
         dismissLabel="dismiss"
         headerText="header"
       >
@@ -45,36 +84,47 @@ describe('EzModal', () => {
       </EzModal>
     );
 
-    fireEvent.click(getByText(submitLabel));
+    // Find and click the first node with text corresponding to the submit label
+    component
+      .findWhere(node => node.type() && node.text() === submitLabel)
+      .at(0)
+      .simulate('click');
     expect(clickSpy).toHaveBeenCalled();
   });
 
   it('calls dismiss handler when dismiss button is clicked', () => {
     const clickSpy = jest.fn();
     const dismissLabel = 'dismiss';
-    const {getByText} = fullRender(
+    const component = mount(
       <EzModal
         isOpen
         dismissLabel={dismissLabel}
         submitLabel="submit"
         onDismiss={clickSpy}
+        appElement={null}
         headerText="header"
       >
         Test
       </EzModal>
     );
 
-    fireEvent.click(getByText(dismissLabel));
+    // Find and click the first node with text corresponding to the dismiss label
+    component
+      .findWhere(node => node.type() && node.text() === dismissLabel)
+      .at(0)
+      .simulate('click');
+
     expect(clickSpy).toHaveBeenCalled();
   });
 
   it('passes through the destructive prop to the submit button ', () => {
     const submitLabel = 'submit';
-    const {getByText} = fullRender(
+    const component = mount(
       <EzModal
         isOpen
         submitLabel={submitLabel}
         destructive
+        appElement={null}
         dismissLabel="dismiss"
         headerText="header"
       >
@@ -82,36 +132,50 @@ describe('EzModal', () => {
       </EzModal>
     );
 
-    const submitButton = getByText(submitLabel);
-    expect(submitButton).toHaveStyle(`background-color: ${standard.colors.destructive}`);
+    // Find the button with text submit
+    const submitButton = component.findWhere(
+      node => node.type() && node.is(EzButton) && node.text() === submitLabel
+    );
+    expect(submitButton.prop('destructive')).toBe(true);
   });
 
-  it('disables the submit button when the form is submitting', () => {
+  it('sets the loading prop on submit button when isSubmitting is true', () => {
     const submitLabel = 'submit';
-    const {getByText} = fullRender(
+    const component = mount(
       <EzModal
         isOpen
         isSubmitting
         submitLabel={submitLabel}
+        appElement={null}
         dismissLabel="dismiss"
         headerText="header"
       >
         test
       </EzModal>
     );
-    const submitButton = getByText(submitLabel);
-    expect(submitButton).toHaveAttribute('disabled');
+    const submitButton = component.findWhere(
+      node => node.type() && node.is(EzButton) && node.text() === submitLabel
+    );
+    expect(submitButton.prop('loading')).toBe(true);
   });
 
   it('sets the disabled prop on dismiss button when isSubmitting is true', () => {
-    const dismissLabel = 'dismiss';
-    const {getByText} = fullRender(
-      <EzModal isOpen isSubmitting dismissLabel={dismissLabel} headerText="header">
+    const dismissLabel = 'submit';
+    const component = mount(
+      <EzModal
+        isOpen
+        isSubmitting
+        dismissLabel={dismissLabel}
+        appElement={null}
+        headerText="header"
+      >
         test
       </EzModal>
     );
-    const dismissButton = getByText(dismissLabel);
-    expect(dismissButton).toHaveAttribute('disabled');
+    const dismissButton = component.findWhere(
+      node => node.type() && node.is(EzButton) && node.text() === dismissLabel
+    );
+    expect(dismissButton.prop('disabled')).toBe(true);
   });
 
   /**
@@ -119,24 +183,28 @@ describe('EzModal', () => {
    */
   describe('Accessibility tests', () => {
     it('should meet accessibility guidelines when required labels / text are given', async () => {
-      const {baseElement, unmount} = fullRender(
-        <EzModal isOpen dismissLabel="dismiss" headerText="header">
+      const wrapper = mount(
+        <EzModal isOpen dismissLabel="dismiss" headerText="header" appElement={null}>
           test
         </EzModal>
       );
-      unmount();
-      const actual = await axe(baseElement.innerHTML);
+      const actual = await axe(wrapper.html());
       expect(actual).toHaveNoViolations();
     });
 
     it('should meet accessibility guidelines when all labels / text are given', async () => {
-      const {baseElement, unmount} = fullRender(
-        <EzModal isOpen headerText="Header" submitLabel="submit" dismissLabel="dismiss">
+      const wrapper = mount(
+        <EzModal
+          isOpen
+          headerText="Header"
+          submitLabel="submit"
+          dismissLabel="dismiss"
+          appElement={null}
+        >
           test
         </EzModal>
       );
-      unmount();
-      const actual = await axe(baseElement.innerHTML);
+      const actual = await axe(wrapper.html());
       expect(actual).toHaveNoViolations();
     });
   });